Fixed input == self.attr("placeholder") bug.
In a case where users entered text that was identical to the placeholder
text, the html5support library would mistakenly treat it as though the
text had in fact been created by the html5support library, and
mistakenly mark it as placeholder text.  This was unacceptable for my
use case, so I fixed it.